model_finetune/requirements.txt
torch==2.2.2
torchvision==0.17.2
facenet-pytorch==2.6.0
numpy
pillow
scikit-learn
tensorboard
matplotlib
wandb
google-cloud-storage